pyproject.toml
[tool.poetry]
name = "SIH2025-25080"
version = "0.1.0"
description = ""
authors = ["Aryan Rajpur <av.rajpurkar@gmail.com>"]
readme = "Readme.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
Flask = "^3.0.3"
Flask-Mail = "^0.9.1"
scikit-learn = "^1.5.2"
rank-bm25 = "^0.2.2"
Flask-Cors = "^5.0.0"
python-dotenv = "^1.0.1"
langchain = "^0.3.0"
openai = "1.55.3"
PyPDF2 = "^3.0.1"
docx2txt = "^0.8"
anthropic = "^0.34.2"
langchain-openai = "^0.2.0"
pillow = "^10.4.0"
redis = "^5.0.8"
fuzzywuzzy = "^0.18.0"
toml = "^0.10.2"
Flask-Caching = "^1.11.0"
cachelib = "^0.13.0"
spacy = "^3.7.5"
python-levenshtein = "^0.26.1"
langchain-community = "^0.3.7"
uvicorn = "^0.32.0"
fastapi = "^0.115.5"
python-multipart = "^0.0.17"
gunicorn = "^21.2.0"
# Heavy ML dependencies - commented out to reduce deployment size
# gliner = "^0.2.13"
pymupdf = "^1.24.14"
pymongo = "^4.10.1"
authlib = "^1.3.2"
pandas = "^2.2.3"
# transformers = "^4.47.0"
# torch = "^2.5.1"
# torchvision = "^0.20.1"
sendgrid = "^6.12.4"
python-docx = "^1.1.2"
numpy = "^1.26.0"
cachetools = "^5.5.0"
google-generativeai = "^0.8.3"
matplotlib = "^3.9.3"
grpcio = "^1.68.1"
python-pptx = "^1.0.2"
"pdfminer.six" = ">=20250506"  # Keep updated - newer versions fix HOCRConverter import issues
pdf2image = "^1.17.0"
paddleocr = "^2.7.0.3"
paddlepaddle = "^2.6.0"
camelot-py = {extras = ["cv"], version = "^1.0.9"}
opencv-python = "^4.10.0.84"
tabula-py = "^2.9.0"
# gliner removed to avoid heavy ML dependency (torch/transformers). Re-add if redaction features are required.
requests = "^2.32.5"
faiss-cpu = "^1.12.0"
scipy = "^1.16.3"
neo4j = "^6.0.3"
google-genai = "^1.53.0"



[tool.poetry.scripts]
start = "app:main"
dev = "app:main"
setup = "setup:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
